[minor] Move the EventTarget methods to lib/EventTarget.js
